refactoring: after review
