+++ packages/client/src/components/Layout/style.module.scss
@@ -5,15 +5,6 @@
   display: flex;
   flex-grow: 1;
   font-family: $font-family;
-
-  a {
-    text-decoration: none;
-    color: $font-color;
-  }
-
-  input {
-    outline: none;
-  }
 }
 
 .Container {
@@ -27,11 +18,11 @@
 }
 
 .GameOver {
-  background-image: url('../../../assets/gameover_bg.jpeg');
+  background-image: url('../../assets/gameover_bg.jpeg');
 }
 
 .Default {
-  background-image: url('../../../assets/default_bg.jpg');
+  background-image: url('../../assets/default_bg.jpg');
 }
 
 .Authors {
