+++ packages/client/src/components/UI/ButtonLink/style.module.scss
@@ -1,6 +1,8 @@
 .ButtonLink {
   margin: 0.25rem;
   border-bottom: 2px solid #ffffff;
+  opacity: 1;
+  transition: opacity 0.3s ease-in-out;
 }
 
 .ButtonLink:hover {
