+++ packages/client/src/components/UI/Header/index.tsx
@@ -1,11 +1,8 @@
 import cn from './style.module.scss'
+import { ButtonHTMLAttributes, FC } from 'react'
 
-interface Props {
-  title: string
-}
-
-const Header = ({ title }: Props) => {
-  return <h1 className={cn.Header}>{title}</h1>
+const Header: FC<ButtonHTMLAttributes<HTMLHeadElement>> = ({ children }) => {
+  return <h1 className={cn.Header}>{children}</h1>
 }
 
 export default Header
