+++ packages/client/src/styles/reset.css
@@ -47,6 +47,10 @@ ol[class] {
 a:not([class]) {
   text-decoration-skip-ink: auto;
 }
+a {
+  color: inherit;
+  text-decoration: none;
+}
 
 /* Упрощаем работу с изображениями */
 img {
@@ -65,6 +69,7 @@ button,
 textarea,
 select {
   font: inherit;
+  outline: none;
 }
 
 /* Удаляем все анимации и переходы для людей, которые предпочитай их не использовать */
