+++ packages/client/src/views/Auth/style.module.scss
@@ -15,8 +15,5 @@
 
 .Message {
   color: $font-color;
-}
-
-.Message {
   padding: 1rem 0.15rem;
 }
